Added external notrivial script
There was a false positive for a file whose only message changed in a
line after a newline, so the lines starting with [+-]" cannot be
excluded. In order to exclude metadata it gets more complicated, so the
script is now external to the makefile.

Using a symlink now to get to the pgtr directory so that the Makefile
doesn't need adjustments.